Set up the project
Generate a package.json file.

Install webpack to the node_modules directory.

Because node_modules folder can become too big,
add a .gitignore file, so that we don't have to
sync the contents of node_modules to github.

Create a src directory with an index.js file.

Create a dist directory with an index.html file.

Create a webpack.config.js.