Set up the site to use tabbed browsing
Put the contents of each tab inside of it's
own module. Each module will export a function
that creates a div element, then appends it to
the DOM.

Write the tab-switching logic inside of index.js.
You should have event listeners for each tab that
wipes out the current contents and then runs the
correct 'tab module' to populate it again.